Added docu for gnuplot and plotutils visualizers.
