+++ graph/docu/docu_04graph_03plotutils.md
@@ -1,0 +1,1 @@
+![](https://raw.githubusercontent.com/dkorolev/Bricks/png/graph/golden/love.png)
