+++ graph/docu/docu_04graph_06gnuplot.md
@@ -1,0 +1,1 @@
+![](https://raw.githubusercontent.com/dkorolev/Bricks/png/graph/golden/gnuplot.png)
