+++ net/api/docu/client/docu_02httpclient_00header.md
@@ -1,6 +1,6 @@
-## HTTP
+## REST API Toolkit
 
 The [`#include "Bricks/net/api/api.h"`](https://github.com/KnowSheet/Bricks/blob/master/net/api/api.h) header enables to run the code snippets below.
 
-### Client
+### HTTP Client
